Revert "feat: migrate data_use_translation.spec"
This reverts commit beefd253dd06fb0eaae3694ed7554f342e6e9589.
